move : CouponStatus 분리
