+++ src/main/java/bc1/gream/domain/common/facade/ChangingCouponStatusFacade.java
@@ -2,7 +2,7 @@ package bc1.gream.domain.common.facade;
 
 import bc1.gream.domain.order.entity.Buy;
 import bc1.gream.domain.order.service.BuyService;
-import bc1.gream.domain.user.entity.CouponStatus;
+import bc1.gream.domain.user.coupon.entity.CouponStatus;
 import bc1.gream.domain.user.entity.User;
 import bc1.gream.domain.user.service.CouponService;
 import lombok.RequiredArgsConstructor;
