+++ src/main/java/bc1/gream/domain/user/coupon/entity/CouponStatus.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.user.entity;
+package bc1.gream.domain.user.coupon.entity;
 
 public enum CouponStatus {
     AVAILABLE,
